Generate unique marker key for consistent renders
Using index as a key for the marker can cause issues (especially on maps with many markers) because the markers are created/updated async. This might result in a changing order of the markers in the locations array between renders.

Using the index as key is discouraged (https://legacy.reactjs.org/docs/lists-and-keys.html)

This fix replaces the key with a value that is specific to the marker.
Ideally, I'd use the GUID of the underlying data object, but that requires much more changes.